+++ assets/Components/SideMenu/SideMenu.jsx
@@ -12,10 +12,11 @@ import ListItem from '@mui/material/ListItem';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import { AccordionDetails, AccordionSummary, Link, Typography } from '@mui/material';
+import { AccordionDetails, AccordionSummary, Box, Link, Typography } from '@mui/material';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import commonStyles from '../../styles/commonStyles.scss';
 import Icon from '../Icon/Icon';
+import configuration from '../../controllerConfig/configuration';
 
 const drawerWidth = 292;
 
@@ -40,7 +41,7 @@ const closedMixin = (theme) => ({
   },
 });
 
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(({ theme, open }) => ({
+const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' && prop !== 'onClose' })(({ theme, open }) => ({
   width: drawerWidth,
   flexShrink: 0,
   whiteSpace: 'nowrap',
@@ -125,7 +126,7 @@ const getMappedMenuItems = (items) => {
                   backgroundColor: 'rgba(0, 47, 54, 0.5)',
                 },
               }}
-              expandIcon={<ArrowDropDownIcon color='secondary' />}
+              expandIcon={<ArrowDropDownIcon color="secondary" />}
             >
               <ListItemIcon>
                 {React.createElement(item.icon, {
@@ -216,29 +217,23 @@ const SideMenu = ({ showSideMenu, toggleSideMenu, enableSideMenu, menuItems }) =
           },
         }}
         open={showSideMenu}
-        onClose={() => toggleSideMenu(false)}
         variant="permanent"
       >
-        <Toolbar sx={{ pl: 2 }} className={`${commonStyles.flex} ${commonStyles.justifyCenter}`}>
-          <Link href={AdServer.LandingUrl} target="_self">
-            <Icon
-              name="logo"
-              sx={{
-                width: 32,
-                height: 32,
-                color: 'secondary.main',
-              }}
-            />
-            {showSideMenu && (
-              <Icon
-                name="adSharesText"
-                sx={{
-                  ml: 1,
-                  mb: 1,
-                  width: 79,
-                  height: 9,
-                  color: 'secondary.main',
-                }}
+        <Toolbar sx={{ pl: 2, maxHeight: '64px' }}>
+          <Link
+            href={AdServer.LandingUrl}
+            target="_self"
+            sx={{ height: '100%', width: '100%' }}
+            className={`${commonStyles.flex} ${commonStyles.justifyCenter} ${commonStyles.alignCenter}`}
+          >
+            {showSideMenu ? (
+              <Box component="img" src={`${configuration.basePath}/assets/panel/LogoH30?${Date.now()}`} maxHeight="100%" maxWidth="100%" />
+            ) : (
+              <Box
+                component="img"
+                src={`${configuration.basePath}/assets/panel/SimpleLogoH30?${Date.now()}`}
+                maxHeight="100%"
+                maxWidth="100%"
               />
             )}
           </Link>