+++ src/Entity/Enum/PanelAssetConfig.php
@@ -11,6 +11,7 @@ enum PanelAssetConfig implements ConfigEnum
     case Favicon48x48;
     case Favicon96x96;
     case LogoH30;
+    case SimpleLogoH30;
 
     public function getModule(): string
     {
@@ -25,6 +26,7 @@ enum PanelAssetConfig implements ConfigEnum
             self::Favicon48x48 => '/favicon-48x48.png',
             self::Favicon96x96 => '/favicon-96x96.png',
             self::LogoH30 => '/assets/images/logo.png',
+            self::SimpleLogoH30 => '/assets/images/simpleLogo.png',
         };
     }
 }