+++ deploy/reload.sh
@@ -3,3 +3,4 @@
 mkdir -p public/panel-assets
 cp -fr public/panel-assets-default/* public/panel-assets
 find var/panel-assets -name "favicon-*" -exec cp {} public/panel-assets \;
+find var/panel-assets -name "*.css" -exec cp {} public/panel-assets/custom.css \;