+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Turnover monitoring
+- Turnover and token flow monitoring
 - Link to wallet on list
 
 ## [1.2.8] - 2023-02-14