+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Turnover monitoring
 - Link to wallet on list
 - Ads.txt configuration
+- Favicons now also change on AdServer and AdController
+- Custom css now also change on AdController
+- Logo css now also change on AdController
+- Opportunity to add simple logo
 
 ## [1.2.8] - 2023-02-14
 ### Fixed