+++ CHANGELOG.md
@@ -5,12 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2023-05-15
 ### Added
 - Turnover and token flow monitoring
 - Dark mode support
 - Turnover monitoring
 - Link to wallet on list
 - Ads.txt configuration
+- Custom CSS support
+- Custom logos support
 
 ## [1.2.8] - 2023-02-14
 ### Fixed
@@ -114,7 +118,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Installer
 - Project setup
 
-[Unreleased]: https://github.com/adshares/adcontroller/compare/v1.2.8...develop
+[Unreleased]: https://github.com/adshares/adcontroller/compare/v1.3.0...develop
+[1.3.0]: https://github.com/adshares/adcontroller/compare/v1.2.8...v1.3.0
 [1.2.8]: https://github.com/adshares/adcontroller/compare/v1.2.7...v1.2.8
 [1.2.7]: https://github.com/adshares/adcontroller/compare/v1.2.6...v1.2.7
 [1.2.6]: https://github.com/adshares/adcontroller/compare/v1.2.5...v1.2.6